+++ test/get-items-between.ts
@@ -1,0 +1,14 @@
+import test from 'ava';
+import getItemsBetween from '../source/helpers/get-items-between';
+
+test('getItemsBetween', t => {
+	const list = ['❤️', '💛', '💚', '💙'];
+
+	t.deepEqual(getItemsBetween<string>(list, '💛', '💚'), ['💛', '💚']);
+	t.deepEqual(getItemsBetween<string>(list, '💚', '💛'), ['💛', '💚']);
+	t.deepEqual(getItemsBetween<string>(list, '❤️', '💙'), ['❤️', '💛', '💚', '💙']);
+	t.deepEqual(getItemsBetween<string>(list, '💙', '❤️'), ['❤️', '💛', '💚', '💙']);
+	t.deepEqual(getItemsBetween<string>(list, undefined, '❤️'), ['❤️']);
+	t.deepEqual(getItemsBetween<string>(list, undefined, '💚'), ['❤️', '💛', '💚']);
+	t.deepEqual(getItemsBetween<string>(list, undefined, '💙'), ['❤️', '💛', '💚', '💙']);
+});
